Remove the checks for dax file existence in examples
This means that an exception will be thrown when we try to run them and
we are missing the files. This is better than an error message because
it ensures that the tests fail if the files are missing.

Unfortunately it results in a confusing error message, so add a check in
Parameters class to make sure that the file exists before it is used.